feat(vocab): merge Words+Practice into single /vocabulary page
- one /vocabulary replaces /words + /practice (legacy redirects)
- word row: 5-step SRS progress bar + relative next-due ("Now", "In 3 days")
- list header: title, "N words" count, spaced-repetition hint + column labels
- book source shown as "📖 from {title}" in word cell
- remove duplicate context sentence (expanded panel only)
- header: drop Stats, Library → Catalog
- user menu: Highlights only
- "Saved to vocabulary" inline status in reader popup
- practice.spec.ts: updated selectors + legacy redirect tests
