+++ apps/web/src/locales/uk.json
@@ -86,7 +86,7 @@
       "known": "Знаю",
       "ignore": "Ігнорувати",
       "save": "Зберегти",
-      "savedStatus": "Збережено",
+      "savedStatus": "Збережено у словник",
       "remove": "Видалити",
       "saving": "Збереження...",
       "loading": "...",
@@ -140,10 +140,6 @@
   },
   "nav": {
     "catalog": "Каталог",
-    "library": "Бібліотека",
-    "words": "Слова",
-    "practice": "Практика",
-    "stats": "Статистика",
     "vocabulary": "Словник",
     "highlights": "Виділення",
     "about": "Про нас",
@@ -509,6 +505,25 @@
       "3": "Контекст",
       "4": "Засвоєно"
     },
+    "due": {
+      "now": "Зараз",
+      "today": "Сьогодні",
+      "tomorrow": "Завтра",
+      "inDays": "Через {n} д.",
+      "inWeeks": "Через {n} тиж.",
+      "inMonths": "Через {n} міс."
+    },
+    "interval": "Наступний інтервал: {n} д.",
+    "listTitle": "Мій словник",
+    "listCount": "{n} слів",
+    "srsHint": "Інтервальне повторення — кожне слово повертається, коли настає час повторення.",
+    "columns": {
+      "word": "Слово",
+      "translation": "Переклад",
+      "progress": "Прогрес",
+      "nextDue": "Коли повторити"
+    },
+    "fromBook": "з «{title}»",
     "filters": {
       "all": "Всі",
       "new": "Нові",
